Fix: Correct API route path and update client for chat functionality. Chat successfully works locally
- Moves chat API handler to  to enable Next.js App Router to correctly resolve the  endpoint. This resolves the 404 error.
- Updates  to call  (was ).
- Consolidates  from  into primary  and updates  import.
- Deletes unused .
- Updates  and  for App Router compatibility.
- Updates task file to reflect these adjustments.
