+++ src/app/api/chat/route.ts
