+++ task-management/integrate-chat-functionality.json
@@ -18,8 +18,7 @@
             "src/components/New/ui/button.tsx",
             "src/components/New/ui/textarea.tsx",
             "src/lib/utils.ts",
-            "src/lib/new-lib/utils.ts",
-            "app/api/ai/chat/route.ts",
+            "src/api/chat/route.ts",
             "src/schemas/generated/*",
             "src/schemas/*",
             "src/utils/validateInput.ts"
@@ -30,9 +29,9 @@
         "[x] Delete placeholder src/components/ChatInterface.tsx and remove its import/usage from pages/index.tsx. Replace with new <Chat /> component.",
         "[x] Consolidate Button component: keep src/components/ui/button.tsx as single source of truth, merging incoming variant props if different; update imports in new chat files.",
         "[x] Consolidate Textarea component the same way.",
-        "[ ] Merge additional helpers from src/lib/new-lib/utils.ts into existing src/lib/utils.ts and delete the duplicate path; fix related imports.",
-        "[ ] Ensure app/api/ai/chat/route.ts is in place and exports the handler (no conversion needed under Next.js).",
-        "[ ] Update Chat component fetch calls to use `/api/ai/chat` (absolute relative path).",
+        "[x] Merge additional helpers from src/lib/new-lib/utils.ts into existing src/lib/utils.ts and delete the duplicate path; fix related imports.",
+        "[x] Ensure API route src/api/chat/route.ts is in place and exports handler.",
+        "[x] Update Chat component fetch calls to use `/api/chat`.",
         "[ ] Move all files in src/schemas/generated/ to src/schemas/, add barrel index.ts exporting relevant types; delete example_usage.ts if not referenced.",
         "[ ] Ensure validateInput.ts is used by the new chat flow; if not, integrate validation or remove file.",
         "[ ] Run `next lint` and fix any import/path errors.",
